When installing, consult SoapySDR for installation dir, not system dir.
This fixes an issue when SoapySDR isn't installed in a "default"
system directory (/usr/local) but is instead part of a larger package
system (like HomeBrew), and thus, installed in an alternate
location.

With this fix, the SoapyUHD plugin file should end up in the proper
library directory no matter how or where SoapySDR was installed.